Add files via upload
"This marks the beginning of my exciting journey with Codsoft! 🚀 My Third task as an intern, thoughtfully provided by the Codsoft team, is to 'ATM INTERFACE', in Java. 📊 I'm thrilled to learn, grow, and contribute to amazing projects during this internship. Let's make the most of this opportunity! 💪 #InternshipKickoff #CodsoftIntern #ThirdTask #java"

Task Description:
1.Create a class to represent the ATM machine.
2. Design the user interface for the ATM, including options such as withdrawing, depositing, and
checking the balance.
3. Implement methods for each option, such as withdraw(amount), deposit(amount), and
checkBalance().
4. Create a class to represent the user's bank account, which stores the account balance.
5. Connect the ATM class with the user's bank account class to access and modify the account
balance.
6. Validate user input to ensure it is within acceptable limits (e.g., sufficient balance for withdrawals).
7. Display appropriate messages to the user based on their chosen options and the success or failure
of their transactions.